Confirmation dialog implemented

+++ LoginPOV/src/app/user/userlist.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-
 import { User } from '../_models/index';
-import { AlertService, UserService } from '../_services/index';
+import { AlertService, UserService, ConfirmationDialogService } from '../_services/index';
 
 @Component({
     selector: 'user-list',
@@ -11,8 +10,7 @@ import { AlertService, UserService } from '../_services/index';
 export class UserListComponent implements OnInit {
     //currentUser: User;
     users: User[] = [];
-
-    constructor(private alertService: AlertService, private userService: UserService) {
+    constructor(private alertService: AlertService, private userService: UserService, private confirmationDialogService: ConfirmationDialogService) {
         //this.currentUser = JSON.parse(localStorage.getItem('currentUser'));
     }
 
@@ -20,35 +18,56 @@ export class UserListComponent implements OnInit {
         this.loadAllUsers();
     }
 
-    deleteUser(id: number) {
-        if (id != JSON.parse(localStorage.getItem('currentUser')).id) {
-            this.userService.delete(id).subscribe(() => { this.loadAllUsers() });
-            this.alertService.success("User deleted");
-        } else {
-            this.alertService.error("Cannot delete oneself");
-        }
-    }
+    // deleteUser(id: number) {
+    //     if (id != JSON.parse(localStorage.getItem('currentUser')).id) {
+    //         this.userService.delete(id).subscribe(() => { this.loadAllUsers() });
+    //         this.alertService.success("User deleted");
+    //     } else {
+    //         this.alertService.error("Cannot delete oneself");
+    //     }
+    // }
 
     editUser(user: any) {
         this.userService.update(user).subscribe(() => { this.loadAllUsers() });
     }
 
     approveUser(user: any) {
-        if (user.id != JSON.parse(localStorage.getItem('currentUser')).id) {
-            this.userService.approve(user).subscribe(() => {
-                this.loadAllUsers(); 
-                if (user.isApproved == false) {
-                    this.alertService.success("User approved");
+        this.confirmationDialogService.confirm('Confirmation', 'Change Approval Status ?')
+        .then((confirmed) => {
+            if (confirmed == true) {
+                if (user.id != JSON.parse(localStorage.getItem('currentUser')).id) {
+                    this.userService.approve(user).subscribe(() => {
+                        this.loadAllUsers(); 
+                        if (user.isApproved == false) {
+                            this.alertService.success("User approved");
+                        } else {
+                            this.alertService.success("User disapproved");
+                        }
+                    });
                 } else {
-                    this.alertService.success("User disapproved");
+                    this.alertService.error("Cannot approve/disapprove oneself");
                 }
-            });
-        } else {
-            this.alertService.error("Cannot approve/disapprove oneself");
-        }
+            }
+        })
+        .catch(() => console.log('User dismissed the dialog (e.g., by using ESC, clicking the cross icon, or clicking outside the dialog)'));
     }
 
     private loadAllUsers() {
         this.userService.getAll().subscribe(users => { this.users = users; });
     }
+
+    public deleteUser(id : number) {
+        this.confirmationDialogService.confirm('Confirmation', 'Confirm Delete ?')
+        .then((confirmed) => {
+            if (confirmed == true) {
+                if (id != JSON.parse(localStorage.getItem('currentUser')).id) {
+                    this.userService.delete(id).subscribe(() => { this.loadAllUsers() });
+                    this.alertService.success("User deleted");
+                } else {
+                    this.alertService.error("Cannot delete oneself");
+                }
+            }
+        })
+        .catch(() => console.log('User dismissed the dialog (e.g., by using ESC, clicking the cross icon, or clicking outside the dialog)'));
+      }
 }
